% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{bm_use_ensembl}
\alias{bm_use_ensembl}
\title{Internal wrapper for BioMart functions}
\usage{
bm_use_ensembl(biomart, dataset, host = NULL)
}
\description{
Internal wrapper for BioMart functions
}
\keyword{internal}
