% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{calculate_orf_usage}
\alias{calculate_orf_usage}
\title{Calculate ORF usage across conditions from a SummarizedExperiment}
\usage{
calculate_orf_usage(sumExp, gene_id)
}
\arguments{
\item{sumExp}{A SummarizedExperiment object containing DOUResults}

\item{gene_id}{A character string specifying the gene ID of interest}
}
\value{
A data.frame with ORF IDs, conditions, and usage estimates
}
\description{
This function extracts post hoc results for a given gene ID and
computes condition-specific ORF usage using emmeans contrasts.
}
\details{
Usage is computed as the inverse logit of the difference between
ribo and RNA strategy emmeans per condition. Only valid emmGrid
objects are used. Strategy levels are assigned using a helper
function \code{\link{assign_strategy_levels}}.
}
\examples{
# Load test data
dir <- system.file("extdata", package = "DOTSeq")
cnt <- read.table(
    file.path(dir, "featureCounts.cell_cycle_subset.txt.gz"),
    header = TRUE,
    comment.char = "#"
)
names(cnt) <- gsub(".*(SRR[0-9]+).*", "\\\\1", names(cnt))

flat <- file.path(dir, "gencode.v47.orf_flattened_subset.gtf.gz")
bed <- file.path(dir, "gencode.v47.orf_flattened_subset.bed.gz")

meta <- read.table(file.path(dir, "metadata.txt.gz"))
names(meta) <- c("run", "strategy", "replicate", "treatment", "condition")
cond <- meta[meta$treatment == "chx", ]
cond$treatment <- NULL

m <- DOTSeqDataSet(
    count_table = cnt,
    condition_table = cond,
    flattened_gtf = flat,
    bed = bed
)

m$sumExp <- m$sumExp[rowRanges(m$sumExp)$is_kept == TRUE, ]
# Subset only one gene
m$sumExp <- m$sumExp[
    rowData(m$sumExp)$gene_id == "ENSG00000119402.18", ]

m <- DOTSeq(dotseq_dataset = m, modules = "DOU")

usage_df <- calculate_orf_usage(
    m$sumExp,
    gene_id = "ENSG00000119402.18"
)
print(usage_df)

}
