% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateTE.R
\name{calculateTE}
\alias{calculateTE}
\title{Calculate Translational Efficiency and Occupancy Shift}
\usage{
calculateTE(
  normCnts,
  rnaSuffix = ".rna",
  riboSuffix = ".ribo",
  sampleDelim = NULL,
  pseudoCnt = 1e-06
)
}
\arguments{
\item{normCnts}{A numeric matrix or data frame of normalized counts,
where rows correspond to ORFs and columns correspond to samples.
RNA-seq and Ribo-seq reads are distinguished by suffixes.}

\item{rnaSuffix}{Character string indicating the suffix of RNA-seq
samples in the column names (default is ".rna").}

\item{riboSuffix}{Character string indicating the suffix of Ribo-seq
samples in the column names (default is ".ribo").}

\item{sampleDelim}{Delimiter used in sample name (default: NULL).}

\item{pseudoCnt}{Numeric value added to counts to avoid division by zero
(default is 1e-6).}
}
\value{
A list with two elements:
\describe{
\item{absoluteTE}{A matrix of translational efficiency (Ribo / RNA).}
\item{occupancyShift}{A matrix of log2 ratios of Ribo-proportion / RNA-proportion within each gene.}
}
}
\description{
This function computes translational efficiency (TE) and occupancy shift
for a set of normalized RNA-seq and Ribo-seq counts. TE is calculated
as the ratio of Ribo-seq counts to RNA-seq counts. Occupancy
shift is calculated as the log2 ratio of Ribo-seq proportion to RNA-seq
proportion within each gene.
}
\examples{
result <- calculateTE(normCounts)
head(result$absoluteTE)
head(result$occupancyShift)

}
