% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{conditionalFormula}
\alias{conditionalFormula}
\alias{conditionalFormula,DOTSeqDataSet-method}
\alias{conditionalFormula,DESeqDataSet-method}
\title{Access the model formula used in DOTSeq analysis}
\usage{
conditionalFormula(object)

\S4method{conditionalFormula}{DOTSeqDataSet}(object)

\S4method{conditionalFormula}{DESeqDataSet}(object)
}
\arguments{
\item{object}{A \code{DOTSeqDataSet} or \code{DESeqDataSet} object.}
}
\value{
A \code{formula} object.
}
\description{
Retrieves the conditional formula stored in a \code{DOTSeqDataSet}
or \code{DESeqDataSet} object. This formula defines the fixed effects
structure used in GLM / GLMM modeling.
}
\examples{
# Create a minimal DOTSeqDataSet object
library(S4Vectors)
dummy <- new("DOTSeqDataSet",
    formula = ~ condition * strategy,
    specs = ~ condition | strategy,
    interactionResults = S4Vectors::DataFrame(),
    strategyResults = S4Vectors::DataFrame()
)
conditionalFormula(dummy)
}
