% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{interactionResults}
\alias{interactionResults}
\alias{interactionResults,DOTSeqDataSet-method}
\alias{interactionResults,DESeqDataSet-method}
\alias{interactionResults<-}
\alias{interactionResults<-,DOTSeqDataSet-method}
\alias{interactionResults<-,DESeqDataSet-method}
\alias{strategyResults}
\alias{strategyResults,DOTSeqDataSet-method}
\alias{strategyResults<-}
\alias{strategyResults<-,DOTSeqDataSet-method}
\title{Access \code{interactionResults} from post hoc analysis}
\usage{
interactionResults(object)

\S4method{interactionResults}{DOTSeqDataSet}(object)

\S4method{interactionResults}{DESeqDataSet}(object)

interactionResults(object) <- value

\S4method{interactionResults}{DOTSeqDataSet}(object) <- value

\S4method{interactionResults}{DESeqDataSet}(object) <- value

strategyResults(object)

\S4method{strategyResults}{DOTSeqDataSet}(object)

strategyResults(object) <- value

\S4method{strategyResults}{DOTSeqDataSet}(object) <- value
}
\arguments{
\item{object}{A \code{DOTSeqDataSet} object.}

\item{value}{A replacement object (e.g., a \code{DESeqDataSet}).}
}
\value{
A \code{DFrame} or \code{data.frame} containing shrinkage-
adjusted estimates and contrast statistics. It is not row-aligned and
will not be subset when subsetting rows of the object.

A \code{DFrame} or \code{data.frame} containing shrinkage-
adjusted estimates and contrast statistics. It is not row-aligned and
will not be subset when subsetting rows of the object.

A \code{DFrame} or \code{data.frame} containing shrinkage-
adjusted estimates and contrast statistics. It is not row-aligned and
will not be subset when subsetting rows of the object.

A \code{DFrame} or \code{data.frame} containing shrinkage-
adjusted estimates and contrast statistics. It is not row-aligned and
will not be subset when subsetting rows of the object.

A \code{DFrame} or \code{data.frame} containing shrinkage-
adjusted estimates and contrast statistics. It is not row-aligned and
will not be subset when subsetting rows of the object.

A \code{DFrame} or \code{data.frame} containing shrinkage-
adjusted estimates and contrast statistics. It is not row-aligned and
will not be subset when subsetting rows of the object.

A \code{DFrame} or \code{data.frame} containing shrinkage-
adjusted estimates and contrast statistics. It is not row-aligned and
will not be subset when subsetting rows of the object.

A \code{DFrame} or \code{data.frame} containing shrinkage-
adjusted estimates and contrast statistics. It is not row-aligned and
will not be subset when subsetting rows of the object.

A \code{DFrame} or \code{data.frame} containing shrinkage-
adjusted estimates and contrast statistics. It is not row-aligned and
will not be subset when subsetting rows of the object.

A \code{DFrame} or \code{data.frame} containing shrinkage-
adjusted estimates and contrast statistics. It is not row-aligned and
will not be subset when subsetting rows of the object.

A \code{DFrame} or \code{data.frame} containing shrinkage-
adjusted estimates and contrast statistics. It is not row-aligned and
will not be subset when subsetting rows of the object.
}
\description{
Retrieves the contrast results stored in a \code{DOTSeqDataSet} object.
These results may include shrinkage-adjusted estimates and contrast
statistics for interaction-specific contrasts.

Retrieves the contrast results stored in a \code{DOTSeqDataSet} object.
These results may include shrinkage-adjusted estimates and contrast
statistics for strategy-specific contrasts.
}
\examples{
# Create a minimal DESeqDataSet object
library(S4Vectors)
dummy <- new("DESeqDataSet",
    formula = ~ condition * strategy,
    specs = ~ condition | strategy,
    interactionResults = S4Vectors::DataFrame()
)
interactionResults(dummy)
}
