% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{DOTSeqObjects-class}
\alias{DOTSeqObjects-class}
\alias{getDOU}
\alias{getDOU,DOTSeqObjects-method}
\alias{getDOU<-}
\alias{getDOU<-,DOTSeqObjects-method}
\alias{getDTE}
\alias{getDTE,DOTSeqObjects-method}
\alias{getDTE<-}
\alias{getDTE<-,DOTSeqObjects-method}
\alias{show,DOTSeqObjects-method}
\alias{interactionResults,DOTSeqObjects-method}
\title{DOTSeqObjects-class}
\usage{
getDOU(object)

\S4method{getDOU}{DOTSeqObjects}(object)

getDOU(object) <- value

\S4method{getDOU}{DOTSeqObjects}(object) <- value

getDTE(object)

\S4method{getDTE}{DOTSeqObjects}(object)

getDTE(object) <- value

\S4method{getDTE}{DOTSeqObjects}(object) <- value

\S4method{show}{DOTSeqObjects}(object)

\S4method{interactionResults}{DOTSeqObjects}(object)
}
\arguments{
\item{object}{A \code{DOTSeqObjects} object.}

\item{value}{A replacement object (e.g., a \code{DOTSeqDataSet}).}
}
\value{
A \code{DOTSeqDataSet} object containing DOU results.
}
\description{
A wrapper class to store both DOU and DTE results from DOTSeq analysis.
}
\section{Slots}{

\describe{
\item{\code{DOU}}{A \code{DOTSeqDataSet} object.}

\item{\code{DTE}}{A \code{DESeqDataSet} object.}
}}

\examples{
# Create dummy objects
dou <- new("DOTSeqDataSet")
dte <- new("DESeqDataSet")
res <- new("DOTSeqObjects", DOU = dou, DTE = dte)
getDOU(res)
getDTE(res)
}
