% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{DOTSeqResults-class}
\alias{DOTSeqResults-class}
\alias{getDOU}
\alias{getDOU,DOTSeqResults-method}
\alias{getDOU<-}
\alias{getDOU<-,DOTSeqResults-method}
\alias{getDTE}
\alias{getDTE,DOTSeqResults-method}
\alias{getDTE<-}
\alias{getDTE<-,DOTSeqResults-method}
\alias{show,DOTSeqResults-method}
\alias{interactionResults,DOTSeqResults-method}
\title{DOTSeqResults-class}
\usage{
getDOU(object)

\S4method{getDOU}{DOTSeqResults}(object)

getDOU(object) <- value

\S4method{getDOU}{DOTSeqResults}(object) <- value

getDTE(object)

\S4method{getDTE}{DOTSeqResults}(object)

getDTE(object) <- value

\S4method{getDTE}{DOTSeqResults}(object) <- value

\S4method{show}{DOTSeqResults}(object)

\S4method{interactionResults}{DOTSeqResults}(object)
}
\description{
A wrapper class to store both DOU and DTE results from DOTSeq analysis.
}
\section{Slots}{

\describe{
\item{\code{DOU}}{A \code{DOTSeqDataSet} object.}

\item{\code{DTE}}{A \code{DESeqDataSet} object.}
}}

