% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{DOUData-validity}
\alias{DOUData-validity}
\title{Validity method for DOUData}
\arguments{
\item{object}{A \code{\link{DOUData-class}} object.}
}
\value{
TRUE if valid, or a character string describing the error.
}
\description{
Checks that the slots in a DOUData object are correctly populated.
}
