% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{douHeatmap}
\alias{douHeatmap}
\title{Run DOU heatmap wrapper}
\usage{
douHeatmap(
  results,
  orf_type,
  species_dataset,
  symbol_col,
  output_file,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{results}{A data frame containing estimates, labels, and groupID}

\item{species_dataset}{A string specifying the Ensembl dataset, e.g., "hsapiens_gene_ensembl".}

\item{symbol_col}{A string indicating the column name for gene symbols in Ensembl,
e.g., "hgnc_symbol" for human or "mgi_symbol" for mouse.}

\item{output_file}{PDF file name}
}
\description{
Run DOU heatmap wrapper
}
