% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{emmSpecs}
\alias{emmSpecs}
\alias{emmSpecs,DOTSeqDataSet-method}
\alias{emmSpecs,DESeqDataSet-method}
\title{Access the emmeans specification formula}
\usage{
emmSpecs(object)

\S4method{emmSpecs}{DOTSeqDataSet}(object)

\S4method{emmSpecs}{DESeqDataSet}(object)
}
\arguments{
\item{object}{A \code{DESeqDataSet} object.}
}
\value{
A \code{formula} object.

A \code{formula} object.

A \code{formula} object.
}
\description{
Retrieves the \code{\link[emmeans]{emmeans}} specification formula
used for post hoc contrast generation.
}
\examples{
# Create a minimal DESeqDataSet object
library(S4Vectors)
dummy <- new("DESeqDataSet",
    formula = ~ condition * strategy,
    specs = ~ condition | strategy,
    interactionResults = S4Vectors::DataFrame(),
    strategyResults = S4Vectors::DataFrame()
)
emmSpecs(dummy)
}
