% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{fmla}
\alias{fmla}
\alias{fmla,DOUData-method}
\alias{fmla,DTEData-method}
\title{Access the model formula used in \pkg{DOTSeq} analysis}
\usage{
fmla(object)

\S4method{fmla}{DOUData}(object)

\S4method{fmla}{DTEData}(object)
}
\arguments{
\item{object}{A \code{\link{DOUData-class}} or
\code{\link{DTEData-class}} object.}
}
\value{
A \code{formula} object.
}
\description{
Retrieves the conditional formula stored in a
\code{\link{DOUData-class}} or \code{\link{DTEData-class}} object.
This formula defines the fixed effects structure used in GLM /
GLMM modeling.
}
\keyword{internal}
