% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{fmla}
\alias{fmla}
\alias{fmla,DOTSeqDataSet-method}
\alias{fmla,DESeqDataSet-method}
\title{Access the model formula used in DOTSeq analysis}
\usage{
fmla(object)

\S4method{fmla}{DOTSeqDataSet}(object)

\S4method{fmla}{DESeqDataSet}(object)
}
\arguments{
\item{object}{A \code{DESeqDataSet} object.}
}
\value{
A \code{formula} object.

A \code{formula} object.

A \code{formula} object.
}
\description{
Retrieves the conditional formula stored in a \code{DOTSeqDataSet}
or \code{DESeqDataSet} object. This formula defines the fixed effects
structure used in GLM / GLMM modeling.
}
\examples{
# Create a minimal DESeqDataSet object
library(S4Vectors)
dummy <- new("DESeqDataSet",
    formula = ~ condition * strategy,
    specs = ~ condition | strategy,
    interactionResults = S4Vectors::DataFrame(),
    strategyResults = S4Vectors::DataFrame()
)
fmla(dummy)
}
