% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{interactionResults}
\alias{interactionResults}
\alias{interactionResults,DOTSeqDataSet-method}
\alias{interactionResults,DESeqDataSet-method}
\alias{interactionResults<-}
\alias{interactionResults<-,DOTSeqDataSet-method}
\alias{interactionResults<-,DESeqDataSet-method}
\title{Access `interactionResults``` from post hoc analysis}
\usage{
interactionResults(object)

\S4method{interactionResults}{DOTSeqDataSet}(object)

\S4method{interactionResults}{DESeqDataSet}(object)

interactionResults(object) <- value

\S4method{interactionResults}{DOTSeqDataSet}(object) <- value

\S4method{interactionResults}{DESeqDataSet}(object) <- value
}
\arguments{
\item{object}{A \code{DOTSeqDataSet} object.}
}
\value{
A \code{DFrame} or \code{data.frame} containing shrinkage-
adjusted estimates and contrast statistics. It is not row-aligned and
will not be subset when subsetting rows of the object.
}
\description{
Retrieves the contrast results stored in a \code{DOTSeqDataSet} object.
These results may include shrinkage-adjusted estimates and contrast
statistics
}
\examples{
# Create a minimal DOTSeqDataSet object
library(S4Vectors)
dummy <- new("DOTSeqDataSet",
    formula = ~ condition * strategy,
    specs = ~ condition | strategy,
    interactionResults = S4Vectors::DataFrame(),
    strategyResults = S4Vectors::DataFrame()
)
interactionResults(dummy)
# Create a minimal DESeqDataSet object
library(S4Vectors)
dummy <- new("DESeqDataSet",
    formula = ~ condition * strategy,
    specs = ~ condition | strategy,
    interactionResults = S4Vectors::DataFrame(),
    strategyResults = S4Vectors::DataFrame()
)
interactionResults(dummy)
}
