% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{lastExonEndPerGroup}
\alias{lastExonEndPerGroup}
\title{Get last end per granges group}
\usage{
lastExonEndPerGroup(grl, keep.names = TRUE)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}}}

\item{keep.names}{a boolean, keep names or not, default: (TRUE)}
}
\value{
a Rle(keep.names = T), or integer vector(F)
}
\description{
Get last end per granges group
}
\examples{
\dontrun{
gr_plus <- GRanges(seqnames = c("chr1", "chr1"),
                   ranges = IRanges(c(7, 14), width = 3),
                   strand = c("+", "+"))
gr_minus <- GRanges(seqnames = c("chr2", "chr2"),
                    ranges = IRanges(c(4, 1), c(9, 3)),
                    strand = c("-", "-"))
grl <- GRangesList(tx1 = gr_plus, tx2 = gr_minus)
lastExonEndPerGroup(grl)
}
}
\author{
Haakon Tjeldnes et al.
}
