% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{numExonsPerGroup}
\alias{numExonsPerGroup}
\title{Get list of the number of exons per group}
\usage{
numExonsPerGroup(grl, keep.names = TRUE)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}}}

\item{keep.names}{a logical, keep names or not, default: (TRUE)}
}
\value{
an integer vector of counts
}
\description{
Can also be used generaly to get number of GRanges object
per GRangesList group
}
\author{
Haakon Tjeldnes et al.
}
