% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDOTHeatmap.R
\name{prepHeatMapData}
\alias{prepHeatMapData}
\title{Prepare data and plot DOT heatmap}
\usage{
prepHeatMapData(results, sigRes, species_dataset, symbol_col)
}
\arguments{
\item{results}{A data frame containing estimates, labels, and groupID.}

\item{sigRes}{A data frame containing significant results.}

\item{species_dataset}{string, e.g. "hsapiens_gene_ensembl"}

\item{symbol_col}{Name of the gene symbol column in Ensembl for your species,
e.g. "hgnc_symbol" for human, "mgi_symbol" for mouse}

\item{output_file}{PDF file name for saving the heatmap.}
}
\description{
Prepare data and plot DOT heatmap
}
