% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDOTHeatmap.R
\name{runDotHeatmap}
\alias{runDotHeatmap}
\title{Run DOT heatmap wrapper}
\usage{
runDotHeatmap(results, sigRes, species_dataset, symbol_col, output_file)
}
\arguments{
\item{results}{A data frame containing estimates, labels, and groupID}

\item{sigRes}{A data frame containing significant results}

\item{species_dataset}{Ensembl dataset string, e.g. "hsapiens_gene_ensembl"}

\item{symbol_col}{Name of the gene symbol column in Ensembl for your species,
e.g. "hgnc_symbol" for human, "mgi_symbol" for mouse}

\item{output_file}{PDF file name}
}
\description{
Run DOT heatmap wrapper
}
