% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{specs}
\alias{specs}
\alias{specs,DOUData-method}
\alias{specs,DTEData-method}
\title{Access the emmeans specification formula}
\usage{
specs(object)

\S4method{specs}{DOUData}(object)

\S4method{specs}{DTEData}(object)
}
\arguments{
\item{object}{A \code{\link{DOUData-class}} or \code{\link{DTEData-class}} object.}
}
\value{
A \code{formula} object.
}
\description{
Retrieves the \code{\link[emmeans]{emmeans}} specification formula
used for post hoc contrast generation.
}
\keyword{internal}
