% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{strandPerGroup}
\alias{strandPerGroup}
\title{Get list of strands per granges group}
\usage{
strandPerGroup(grl, keep.names = TRUE)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}}}

\item{keep.names}{a boolean, keep names or not, default: (TRUE)}
}
\value{
a vector named/unnamed of characters
}
\description{
Get list of strands per granges group
}
\references{
Tjeldnes, H., Labun, K., Torres Cleuren, Y. et al.
ORFik: a comprehensive R toolkit for the analysis of translation.
BMC Bioinformatics 22, 336 (2021). DOI: 10.1186/s12859-021-04254-w
}
\author{
Haakon Tjeldnes et al.
}
\keyword{internal}
