% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{strategyResults}
\alias{strategyResults}
\alias{strategyResults,DOTSeqDataSet-method}
\alias{strategyResults,DESeqDataSet-method}
\alias{strategyResults<-}
\alias{strategyResults<-,DOTSeqDataSet-method}
\alias{strategyResults<-,DESeqDataSet-method}
\title{Access `strategyResults``` from post hoc analysis}
\usage{
strategyResults(object)

\S4method{strategyResults}{DOTSeqDataSet}(object)

\S4method{strategyResults}{DESeqDataSet}(object)

strategyResults(object) <- value

\S4method{strategyResults}{DOTSeqDataSet}(object) <- value

\S4method{strategyResults}{DESeqDataSet}(object) <- value
}
\arguments{
\item{object}{A \code{DOTSeqDataSet} object.}
}
\value{
A \code{DFrame} or \code{data.frame} containing shrinkage-
adjusted estimates and contrast statistics. It is not row-aligned and
will not be subset when subsetting rows of the object.
}
\description{
Retrieves the contrast results stored in a \code{DOTSeqDataSet} object.
These results may include shrinkage-adjusted estimates and contrast
statistics
}
\examples{
# Create a minimal DOTSeqDataSet object
library(S4Vectors)
dummy <- new("DOTSeqDataSet",
    formula = ~ condition * strategy,
    specs = ~ condition | strategy,
    interactionResults = S4Vectors::DataFrame(),
    strategyResults = S4Vectors::DataFrame()
)
strategyResults(dummy)
# Create a minimal DESeqDataSet object
library(S4Vectors)
dummy <- new("DESeqDataSet",
    formula = ~ condition * strategy,
    specs = ~ condition | strategy,
    interactionResults = S4Vectors::DataFrame(),
    strategyResults = S4Vectors::DataFrame()
)
strategyResults(dummy)
}
