% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getExonicReads}
\alias{getExonicReads}
\title{Filter BAM files to retain only reads overlapping exonic regions}
\usage{
getExonicReads(gr, bam_files, coding_genes_only = TRUE, verbose = TRUE)
}
\arguments{
\item{gr}{A \code{GRanges} object with a \code{TxDb} object stored
in its metadata slot under \code{metadata(gr)$txdb}.}

\item{bam_files}{A character vector of paths to BAM files to be
filtered.}

\item{coding_genes_only}{Logical; if \code{TRUE}, restrict filtering
to coding genes only (i.e., genes with CDS).}

\item{verbose}{Logical; if \code{TRUE}, print progress and runtime
messages.}
}
\value{
This function does not return a value. It creates filtered
and sorted BAM files for each input BAM.
}
\description{
This function filters BAM files to retain reads overlapping exonic
regions #' defined in a TxDb object stored in the metadata of a
GRanges object. Optionally, it restricts to coding genes only.
Filtered BAMs are sorted and saved with \code{.exonic.sorted.bam} suffixes.
}
\examples{
\dontshow{if (requireNamespace("TxDb.Dmelanogaster.UCSC.dm3.ensGene", quietly = TRUE) && requireNamespace("pasillaBamSubset", quietly = TRUE)) withAutoprint(\{ # examplesIf}
library(TxDb.Dmelanogaster.UCSC.dm3.ensGene)
library(pasillaBamSubset)
library(AnnotationDbi)
library(GenomeInfoDb)

# Save a subset of TxDb as an sqlite file
txdb_chr4 <- keepSeqlevels(
    TxDb.Dmelanogaster.UCSC.dm3.ensGene, 
    "chr4", 
    pruning.mode = "coarse"
)
txdb_path <- file.path(getwd(), "dm3_chr4.sqlite")
saveDb(txdb_chr4, file = txdb_path)

# Create a GRanges object with a link to the TxDb sqlite file, 
# which is required for getExonicReads()
gr <- GRanges(seqnames = "chr4", ranges = IRanges(start = 233, end = 2300))
metadata(gr)$txdb <- txdb_path

getExonicReads(gr, bam_files = c(untreated1_chr4()))

# Get the directory of the BAM file and clean up output files
bam_dir <- dirname(untreated1_chr4())
output_files <- list.files(
    path = bam_dir,
    pattern = "*exonic.*",
    full.names = TRUE
)
file.remove(c(txdb_path, output_files))
\dontshow{\}) # examplesIf}
}
