% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{pairedData}
\alias{pairedData}
\title{Prepare Data and Plot Heatmap for DOU}
\usage{
pairedData(results, orf_type, species_dataset, symbol_col, threshold = 0.05)
}
\arguments{
\item{results}{An output data frame from testDOT containing estimates, gene labels, and groupID.}

\item{orf_type}{Character string specifying the ORF type to compare with mORF.
Accepts either "uORF" or "dORF".}

\item{species_dataset}{A string specifying the Ensembl dataset, e.g., "hsapiens_gene_ensembl".}

\item{symbol_col}{A string indicating the column name for gene symbols in Ensembl,
e.g., "hgnc_symbol" for human or "mgi_symbol" for mouse.}

\item{threshold}{Numeric value specifying the empirical FDR threshold (default = 0.05).}
}
\description{
This function generates a heatmap for DOU (Differential ORF Usage) using
estimates from a results data frame, comparing mORFs with either uORFs or dORFs.
}
