% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDOTHeatmap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot DOT heatmap}
\usage{
plotHeatmap(
  orderedMatrix,
  rowDendClean,
  highlight_df,
  gene_labels,
  color_palette,
  color_breaks,
  abs_max,
  width = NULL,
  height = NULL,
  output_file = "dot.pdf"
)
}
\arguments{
\item{orderedMatrix}{matrix of values to plot}

\item{rowDendClean}{dendrogram for row ordering}

\item{highlight_df}{dataframe for highlighting tiles}

\item{gene_labels}{vector of gene labels}

\item{color_palette}{vector of colors}

\item{color_breaks}{breaks for colors}

\item{abs_max}{maximum absolute value}

\item{output_file}{filename for PDF output}
}
\description{
Plot DOT heatmap
}
