% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{reverseMinusStrandPerGroup}
\alias{reverseMinusStrandPerGroup}
\title{Reverse minus strand}
\usage{
reverseMinusStrandPerGroup(grl, onlyIfIncreasing = TRUE)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}}}

\item{onlyIfIncreasing}{logical, default (TRUE), only reverse if decreasing}
}
\value{
a \code{\link[GenomicRanges]{GRangesList}}
}
\description{
Reverse minus strand per group in a GRangesList
Only reverse if minus strand is in increasing order
}
\references{
Tjeldnes, H., Labun, K., Torres Cleuren, Y. et al.
ORFik: a comprehensive R toolkit for the analysis of translation.
BMC Bioinformatics 22, 336 (2021). DOI: 10.1186/s12859-021-04254-w
}
\author{
Haakon Tjeldnes et al.
}
\keyword{internal}
